utils/pyproject.toml
[tool.poetry]
name = "intake-formats"
version = "0.1.0"
description = "Sekoia.io Intake Formats"
authors = []

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.4"
requests = "^2.31.0"
PyYAML = "^6.0"
pytest-clarity = "^1.0.1"
cookiecutter = "^2.4.0"
typer = "^0.9.4"
orjson = "^3.10.3"
pydantic = "^2.7.1"
pillow = "^10.3.0"
rich = "^13.7.1"
structlog = "^24.1.0"
poethepoet = "^0.26.1"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
isort = "^5.10.1"
black = { version = "^24.3.0", extras = ["colorama"] }

[tool.black]
line-length = 119
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 119

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks.lint]
sequence = [
    { cmd = "find . -type f -name '*.py' ! -name '__init__.py' ! -name '__main__.py' -exec poetry run autoflake --in-place --remove-all-unused-imports --remove-duplicate-keys --remove-unused-variables {} +" },
    { cmd = "poetry run black ." },
    { cmd = "poetry run isort ." },
]
help = "Make code linting and formatting."
